Create IMU Sensor-LMS9DS1
9-axis iNEMO inertial module (IMU)
The INEMO is the 9-axis motion sensing system on board which measures a body's specific force, angular rate and the orientation of the body using a combination of accelerometers, gyroscope and magnetometers.
The LSM9DS1 is a system-in-package featuring a 3D digital linear acceleration sensor, a 3D digital angular rate sensor, and a 3D digital magnetic sensor. 
Accelerometer
An accelerometer is an electromechanical device used to measure acceleration forces. Such forces may be static, like the continuous force of gravity or, as is the case with many mobile devices, dynamic to sense movement or vibrations.
The accelerometer of the IMU can be used to measure the device’s orientation.